refactor: font 최적화
