--- src/styles/text.css.ts
@@ -1,23 +1,0 @@
-import { fontFace } from "@vanilla-extract/css";
-
-export const Pretendard = fontFace([
-  {
-    src: `url("/fonts/Pretendard-Regular.otf")`,
-    fontWeight: "normal",
-  },
-  {
-    src: `url("/fonts/Pretendard-Medium.otf")`,
-    fontWeight: "bold",
-  },
-]);
-
-export const Rubik = fontFace([
-  {
-    src: `url("/fonts/Rubik-Regular.ttf")`,
-    fontWeight: "normal",
-  },
-  {
-    src: `url("/fonts/Rubik-Medium.ttf")`,
-    fontWeight: "bold",
-  },
-]);
